+++ packages/contracts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.3.0
+
+### Minor Changes
+
+- date for request & response
+
 ## 3.2.2
 
 ### Patch Changes